docs on model and minor classes

+++ src/it/matlice/matlichess/view/ScreenLocation.java
@@ -2,19 +2,24 @@ package it.matlice.matlichess.view;
 
 import java.awt.*;
 
+/**
+ * Utility class that adds new methods to {@link Point}
+ */
 public class ScreenLocation extends Point {
 
     public ScreenLocation(int w, int h){
         super(w, h);
     }
 
-    public ScreenLocation(){super();}
+    public ScreenLocation(){
+        super();
+    }
 
     public ScreenLocation diff(Point d){
-        return new ScreenLocation((int) (this.x - d.x), (int)(this.y - d.y));
+        return new ScreenLocation(this.x - d.x, this.y - d.y);
     }
 
     public ScreenLocation diff(int x, int y){
-        return new ScreenLocation((int) (this.x - x), (int)(this.y - y));
+        return new ScreenLocation(this.x - x, this.y - y);
     }
 }
